+++ src/single_window/render.h
@@ -215,12 +215,12 @@ class Render {
 
   float control_bar_pos_x_ = 0;
   float control_bar_pos_y_ = 30;
-  float mouse_diff_control_bar_pos_x_ = 0;
-  float mouse_diff_control_bar_pos_y_ = 0;
-  float mouse_pos_x_ = 0;
-  float mouse_pos_y_ = 0;
-  float mouse_pos_x_last_ = 0;
-  float mouse_pos_y_last_ = 0;
+  int mouse_diff_control_bar_pos_x_ = 0;
+  int mouse_diff_control_bar_pos_y_ = 0;
+  int mouse_pos_x_ = 0;
+  int mouse_pos_y_ = 0;
+  int mouse_pos_x_last_ = 0;
+  int mouse_pos_y_last_ = 0;
 
   int main_window_width_real_ = 720;
   int main_window_height_real_ = 540;
